natal-chart-generation/requirements.txt
# NOTE: Don't auto-generate (e.g. with `pipreqs`)
# Dependencies must be compatible with Python3.7
boto3==1.26.102
numpy==1.20.0
Pillow==9.4.0
pyswisseph==2.10.3.1
pytz==2023.3
requests==2.28.2
disjoint-set==0.7.3
